Implement new algorithm

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/find/dialog/employmentsystem/PegManagementDto.java
@@ -11,7 +11,7 @@ public class PegManagementDto {
 	private GeneralDate usageDate;
 	
 	// field 使用日数
-	private int usageDay;
+	private double usageDay;
 	
 	// field 使用時間数
 	private int usageHour;
